chore: optimize Toast implementation
