+++ apps/mobile/src/components/Toast/ToastProviderWrapper.tsx
@@ -1,0 +1,21 @@
+import React, { ReactNode } from 'react';
+
+import {  ToastContextSetter } from './staticMethods';
+import { ToastProvider as OriginalToastProvider } from './ToastProvider';
+
+// Enhanced ToastProvider that sets global context
+interface ToastProviderProps {
+  children: ReactNode;
+}
+
+export const ToastProvider: React.FC<ToastProviderProps> = ({ children }) => {
+  return (
+    <OriginalToastProvider>
+      <ToastContextSetter>{children}</ToastContextSetter>
+    </OriginalToastProvider>
+  );
+};
+
+
+
+export {staticMethods} from './staticMethods';
